Fixed stopwatch/timer ui glitch for andriod
When navigating to the doing tab (using android app), only the timer buttons are be displayed. This seems to be caused by adding button class to the links.
Changed the timer/stopwatch to follow grid design using rows and columns.
Removed the button class and replaced it with custom timer-btn class that styles the link to resemble buttons.